Moved ppq provider implementation to separate file
For consistency with NanoGPT provider implementation.
